Eliminate a shell injection vulnerability.
This is a shell injection vulnerability. An attacker can put arbitrary text into the parameter, which will be executed by the shell. It would safer to ensure that only safe characters would be considered, and to not invoke the shell at all. By taking both steps, we make it much harder for an attacker to find a way to work around it.